move almost everything into 'src'
