+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'app.dart';
+import 'src/app.dart';
 
 // initialize constants
 final String url = "http://localhost:3000";
