+++ lib/src/models/chart_data.dart
